% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_by_dist.R
\name{bin_by_dist}
\alias{bin_by_dist}
\title{bin_by_dist}
\arguments{
\item{x}{A checked_csv filename that you want to bin}

\item{bin_size}{the size of each bin in metres}
}
\value{
None - Saves a csv file ending in _binned_by_*bin_size*.csv
}
\description{
This function is not yet operational
}
