% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dat_file.R
\name{clean_dat_file}
\alias{clean_dat_file}
\title{Cleans a Raw Dat file}
\usage{
clean_dat_file(x)
}
\arguments{
\item{x}{A filename of a raw .Dat file}
}
\value{
NONE - instead saves two files of tidied data, "*_trimmed.csv" and "_summary.txt"
}
\description{
After  from STISIM and returns two files. 1. a csv file ending in "_trimmed.csv", and a txt
file containing a summary of traffic violations and other notable events
}
