% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_trimmed_data.R
\name{load_trimmed_data}
\alias{load_trimmed_data}
\alias{load_trimmed}
\title{Plots raw trimmed data by time for visual data check}
\usage{
load_trimmed(x)
}
\arguments{
\item{x}{filename}
}
\value{
A dataframe
}
\description{
Plots raw trimmed data by time for visual data check
}
