% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_time.R
\name{plot_by_time}
\alias{plot_by_time}
\title{Plots raw trimmed data by time for visual data check}
\arguments{
\item{x}{A dataframe of the drivesim parameters}
}
\value{
A faceted figure of all variables recorded over distance
}
\description{
Plots raw trimmed data by time for visual data check
}
