% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_checked_data.R
\name{save_checked_data}
\alias{save_checked_data}
\title{Plots raw trimmed data by time for visual data check}
\usage{
save_checked_data(x, file_to_check)
}
\arguments{
\item{x}{A dataframe of the drivesim parameters}

\item{file_to_check}{The filename of the file being checked}
}
\value{
None - Saves a csv of "_checked.csv"
}
\description{
Plots raw trimmed data by time for visual data check
}
